requirements.txt
colorama==0.4.6
glfw==2.7.0
mistune==3.0.1
nbconvert==7.8.0
numpy==1.26.4
pygame==2.5.2
PyGLM==2.7.1
PyOpenGL==3.1.7
PyWavefront==1.3.3
scipy==1.13.0
tqdm==4.66.2
